quotes in expansions might be working norminetted parsing
