+++ incs/libft/ft_matrix_dup.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_matrix_dup.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dpaes-so <dpaes-so@student.42.fr>          +#+  +:+       +#+        */
+/*   By: daniel <daniel@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/12 17:06:49 by dpaes-so          #+#    #+#             */
-/*   Updated: 2025/05/12 17:19:09 by dpaes-so         ###   ########.fr       */
+/*   Updated: 2025/05/21 02:29:39 by daniel           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
