+++ srcs/parser/expansions.c
@@ -3,127 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   expansions.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dpaes-so <dpaes-so@student.42.fr>          +#+  +:+       +#+        */
+/*   By: daniel <daniel@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/11 15:58:38 by dgarcez-          #+#    #+#             */
-/*   Updated: 2025/05/20 19:12:12 by dpaes-so         ###   ########.fr       */
+/*   Updated: 2025/05/21 02:26:43 by daniel           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../incs/mini_header.h"
-
-int	amount_quotes(t_token *token)
-{
-	char	quote;
-	int		count;
-	int		i;
-
-	quote = 0;
-	count = 0;
-	i = 0;
-	while ((*token).value[i])
-	{
-		if ((*token).value[i] == '\'' || (*token).value[i] == '\"')
-		{
-			quote = (*token).value[i];
-			if ((*token).value[i] != '\0' && (*token).value[i] == quote)
-				count++;
-			i++;
-			while ((*token).value[i] != '\0' && (*token).value[i] != quote)
-				i++;
-			if ((*token).value[i] != '\0' && (*token).value[i] == quote)
-			{
-				count++;
-				i++;
-			}
-		}
-		else
-			i++;
-	}
-	return (count);
-}
-
-void	removed(t_token *token, char *temp)
-{
-	char	quote;
-	int		i;
-	int		j;
-
-	j = 0;
-	i = 0;
-	while (temp[i])
-	{
-		if (temp[i] == '\'' || temp[i] == '\"')
-		{
-			quote = temp[i++];
-			while (temp[i] != '\0' && temp[i] != quote)
-			{
-				(*token).value[j] = temp[i];
-				j++;
-				i++;
-			}
-		}
-		else
-		{
-			(*token).value[j] = temp[i];
-			j++;
-		}
-		if (temp[i] != '\0')
-			i++;
-	}
-}
-
-bool	remove_quotes(t_token *token)
-{
-	int		count;
-	char	*temp;
-
-	temp = ft_strdup((*token).value);
-	if (temp == NULL)
-		return (false);
-	count = amount_quotes(token);
-	if ((*token).type == T_HERE_DOC && count > 0)
-		(*token).in_quotes = true;
-	free((*token).value);
-	(*token).value = ft_calloc((ft_strlen(temp) - count) + 1, sizeof(char));
-	if ((*token).value == NULL)
-		return (free(temp), false);
-	(*token).value[ft_strlen(temp) - count] = '\0';
-	removed(token, temp);
-	free(temp);
-	if (count == 0)
-		return (false);
-	return (true);
-}
-
-// int	unclosed_split_quotes(char *value)
-// {
-// 	int		i;
-// 	int		quote;
-// 	char	cquote;
-
-// 	i = 0;
-// 	quote = 0;
-// 	while (value[i])
-// 	{
-// 		if (value[i] == '\"' || value[i] == '\'')
-// 		{
-// 			quote++;
-// 			cquote = value[i];
-// 			i++;
-// 			while (value[i] && value[i] != cquote)
-// 				i++;
-// 			if (value[i] == cquote)
-// 			{
-// 				quote++;
-// 				i++;
-// 			}
-// 		}
-// 		else
-// 			i++;
-// 	}
-// 	return (quote);
-// }
 
 int	new_tokens_amount(t_token *tokens, int i, int j)
 {
@@ -144,7 +31,8 @@ int	new_tokens_amount(t_token *tokens, int i, int j)
 			amount++;
 		while (res && res[j])
 			j++;
-		if (tokens[i].type >= T_HERE_DOC && tokens[i].type  <= T_APPEND_REDIR && j > 1)
+		if (tokens[i].type >= T_HERE_DOC && tokens[i].type <= T_APPEND_REDIR
+			&& j > 1)
 			return (freetrix(res), -1);
 		amount += j;
 		i++;
@@ -189,7 +77,8 @@ t_token	*create_new_tokens(t_token *tokens, int amount, int i, int k)
 		else
 		{
 			new_tokens[k].value = ft_strdup(tokens[i].value);
-			if (tokens[i].type == T_PIPE || (tokens[i].type >= T_HERE_DOC && tokens[i].type <= T_APPEND_REDIR))
+			if (tokens[i].type == T_PIPE || (tokens[i].type >= T_HERE_DOC
+					&& tokens[i].type <= T_APPEND_REDIR))
 				new_tokens[k].type = token_type(new_tokens[k].value, 1);
 			else
 				new_tokens[k].type = token_type(new_tokens[k].value, 0);
