+++ srcs/parser/parsing.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parsing.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dpaes-so <dpaes-so@student.42.fr>          +#+  +:+       +#+        */
+/*   By: daniel <daniel@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/17 18:25:00 by dgarcez-          #+#    #+#             */
-/*   Updated: 2025/05/20 18:06:23 by dpaes-so         ###   ########.fr       */
+/*   Updated: 2025/05/20 23:14:33 by daniel           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 t_tree	*tree_maker(t_token **array)
 {
 	t_tree	*root;
-	root = NULL;
 	int		i;
 
+	root = NULL;
 	i = 0;
 	create_tree(&root, array, false, &i);
 	return (root);
@@ -26,8 +26,8 @@ t_tree	*tree_maker(t_token **array)
 t_token	**create_array(t_token *tokens)
 {
 	t_token	**array;
-	array = NULL;
 
+	array = NULL;
 	array = array_creation(tokens);
 	init_array(array, tokens);
 	return (array);
@@ -37,7 +37,7 @@ t_token	**create_array(t_token *tokens)
 t_tree	*parser(char *input, t_mini *shell)
 {
 	t_token	*tokens;
-	t_token *new_tokens;
+	t_token	*new_tokens;
 	t_token	**array;
 	t_tree	*tree;
 
@@ -49,10 +49,10 @@ t_tree	*parser(char *input, t_mini *shell)
 	if (tokens == NULL)
 		return (freetrix(shell->pipex.path), NULL);
 	if (error_syntax(tokens) == false)
-		return (free_tokens(tokens),freetrix(shell->pipex.path), NULL);
+		return (free_tokens(tokens), freetrix(shell->pipex.path), NULL);
 	new_tokens = expand_strs(tokens, shell);
 	if (new_tokens == NULL)
-		return (free_tokens(tokens),freetrix(shell->pipex.path), NULL);
+		return (free_tokens(tokens), freetrix(shell->pipex.path), NULL);
 	array = create_array(new_tokens);
 	free_tokens(tokens);
 	free_tokens(new_tokens);
