+++ srcs/parser/quotes_funcs.c
@@ -1,0 +1,140 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   quotes_funcs.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: daniel <daniel@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/05/21 00:44:11 by daniel            #+#    #+#             */
+/*   Updated: 2025/05/21 02:10:14 by daniel           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../incs/mini_header.h"
+
+int	amount_quotes(t_token *token)
+{
+	char	quote;
+	int		count;
+	int		i;
+
+	count = 0;
+	i = 0;
+	while ((*token).value[i])
+	{
+		if ((*token).value[i] == '\'' || (*token).value[i] == '\"')
+		{
+			quote = (*token).value[i];
+			count++;
+			i++;
+			while ((*token).value[i] && (*token).value[i] != quote)
+				i++;
+			if ((*token).value[i] == quote)
+			{
+				count++;
+				i++;
+			}
+		}
+		else
+			i++;
+	}
+	return (count);
+}
+
+void	removed(t_token *token, char *temp)
+{
+	char	quote;
+	int		i;
+	int		j;
+
+	j = 0;
+	i = 0;
+	while (temp[i])
+	{
+		if (temp[i] == '\'' || temp[i] == '\"')
+		{
+			quote = temp[i++];
+			while (temp[i] != '\0' && temp[i] != quote)
+			{
+				(*token).value[j++] = temp[i];
+				i++;
+			}
+		}
+		else
+		{
+			(*token).value[j] = temp[i];
+			j++;
+		}
+		if (temp[i] != '\0')
+			i++;
+	}
+}
+
+char	*put_quotes(char *temp, int count)
+{
+	int		i;
+	int		j;
+	char	*new_expand;
+
+	i = 0;
+	j = 0;
+	new_expand = ft_calloc(ft_strlen(temp) + (count * 2) + 1, sizeof(char));
+	if (new_expand == NULL)
+		return (NULL);
+	while (temp[i])
+	{
+		if (temp[i] == '\'')
+		{
+			new_expand[j++] = '\"';
+			new_expand[j++] = '\'';
+			new_expand[j++] = '\"';
+		}
+		else
+			new_expand[j++] = temp[i];
+		i++;
+	}
+	return (new_expand);
+}
+
+char	*add_quotes(char *temp, int *flag)
+{
+	int		i;
+	int		count;
+
+	count = 0;
+	i = 0;
+	while (temp[i])
+	{
+		if (temp[i] == '\'')
+			count++;
+		i++;
+	}
+	if (count == 0)
+		return (temp);
+	temp = put_quotes(temp, count);
+	*flag = 2;
+	return (temp);
+}
+
+bool	remove_quotes(t_token *token)
+{
+	int		count;
+	char	*temp;
+
+	temp = ft_strdup((*token).value);
+	if (temp == NULL)
+		return (false);
+	count = amount_quotes(token);
+	if ((*token).type == T_HERE_DOC && count > 0)
+		(*token).in_quotes = true;
+	free((*token).value);
+	(*token).value = ft_calloc((ft_strlen(temp) - count) + 1, sizeof(char));
+	if ((*token).value == NULL)
+		return (free(temp), false);
+	(*token).value[ft_strlen(temp) - count] = '\0';
+	removed(token, temp);
+	free(temp);
+	if (count == 0)
+		return (false);
+	return (true);
+}
